requirements.txt
pyrogram==2.0.106
TgCrypto
python-dotenv
psutil
